requirements.txt
 Django>=5.1.3
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.8
Pillow>=11.0.0
django-environ>= 0.11.2
django-filter 24.3
